Added internal networks
I noticed, that the different applications could not talk to their
databases. I think this must be, because the main service of that
application was set to be in the traefik_network. I think this way it
was not added to the default network of that Docker Compose project.

I now added explicit networks to all applications.