Improved volumes
- Using volumes for persisten data that does not need to be backed up.
- Using bind mounts for peristent data that does need to be backed up.

This way I can have one single folder with all the important data in it,
that I can just backup using rsync or any other method of copying files.

It not only makes it easier as the important files are already directly
on the disk somehwere. I also circumvent the problem of having too
little space for coping all the data from withtin the docker volume to
disk. That only work if you have at least half the disk free.